Cothread: Fix bug in parsing stack size
Python parsing a size_t as if it were an int doesn't work properly on
64 bit systems, and can produce rather scary stack sizes!

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@67098 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ context/_coroutine.c
@@ -90,7 +90,7 @@ static PyObject * coroutine_create(PyObject *self, PyObject *args)
 {
     struct cocore *parent;
     PyObject *action;
-    size_t stack_size;
+    int stack_size;
     if (PyArg_ParseTuple(args, "O&OI",
             get_cocore, &parent, &action, &stack_size))
     {
